chore: refactor api server out of db folder
