+++ api/Cargo.toml
@@ -1,0 +1,71 @@
+[package]
+edition = "2021"
+license = "MIT"
+name    = "arithmetic-api"
+version = "0.1.0"
+
+[[bin]]
+name = "server"
+path = "src/bin/server.rs"
+
+[target.'cfg(target_os = "linux")'.dependencies]
+io-uring = { version = "0.7.2", optional = true }
+
+[target.'cfg(not(target_env = "msvc"))'.dependencies]
+tikv-jemallocator = { version = "0.6.0", optional = true }
+
+[features]
+default = [ "tikv-jemallocator" ]
+
+[dependencies]
+# Core dependencies
+chrono = { version = "0.4", features = [ "serde" ] }
+serde = { workspace = true }
+thiserror = "1.0"
+tokio = { workspace = true }
+tracing = { workspace = true }
+uuid = { version = "1.0", features = [ "serde", "v4" ] }
+
+# Database and arithmetic dependencies
+alloy-sol-types = { workspace = true }
+arithmetic-db = { path = "../db" }
+arithmetic-lib = { path = "../lib" }
+sindri = { version = "0.3.1", features = [ "sp1-v5" ] }
+
+# API framework dependencies
+async-graphql = { version = "7.0", features = [ "chrono" ] }
+async-graphql-axum = "7.0"
+async-trait = "0.1"
+axum = { version = "0.8", features = [ "ws" ] }
+futures-util = "0.3"
+hex = "0.4"
+http = "1.0"
+hyper = "1.0"
+serde_json = "1.0"
+tokio-stream = "0.1"
+tower = { version = "0.4", features = [ "util" ] }
+tower-http = { version = "0.5", features = [
+    "compression-gzip",
+    "cors",
+    "limit",
+    "timeout",
+    "trace",
+] }
+
+# Server binary dependencies
+clap = { version = "4.0", features = [ "derive" ] }
+dotenv = "0.15"
+tracing-subscriber = { version = "0.3", features = [ "env-filter" ] }
+
+# Middleware dependencies
+rand = "0.8"
+regex = "1.0"
+sha2 = "0.10"
+url = "2.5"
+
+[dev-dependencies]
+axum-test = "17.3.0"
+tempfile = "3.10"
+tracing-subscriber = "0.3"
+tracing-test = "0.2"
+uuid = { version = "1.0", features = [ "v4" ] }
